backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
fastmcp
openai
pydantic
mcp
